Move the uploading code to the serial event uploader; If IOpipe is not enabled then do not increment the bad request count (this is just an old hold over from ancient tests).
