+++ src/main/java/com/iopipe/http/SerialEventUploader.java
@@ -2,6 +2,9 @@ package com.iopipe.http;
 
 import com.iopipe.http.RemoteRequest;
 import com.iopipe.IOpipeEventUploader;
+import java.util.concurrent.atomic.AtomicInteger;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
 
 /**
  * This is an uploader which is completely serial based and it will only
@@ -13,8 +16,16 @@ import com.iopipe.IOpipeEventUploader;
 public final class SerialEventUploader
 	implements IOpipeEventUploader
 {
+	/** Logging. */
+	private static final Logger _LOGGER =
+		LogManager.getLogger(SerialEventUploader.class);
+	
 	/** The connection to the remote service to use. */
 	protected final RemoteConnection connection;
+	
+	/** The number of bad requests. */
+	private final AtomicInteger _badresultcount =
+		new AtomicInteger();
 	
 	/**
 	 * Initializes the serialized event uploader.
@@ -38,7 +49,7 @@ public final class SerialEventUploader
 	 */
 	public final int badRequestCount()
 	{
-		throw new Error("TODO");
+		return _badresultcount.get();
 	}
 	
 	/**
@@ -52,7 +63,60 @@ public final class SerialEventUploader
 		if (__r == null)
 			throw new NullPointerException();
 		
-		throw new Error("TODO");
+		// Generate report
+		try
+		{
+			// Report what is to be sent
+			_LOGGER.debug(() -> "Send: " + __r + " " + __debugBody(__r));
+			
+			RemoteResult result = this.connection.send(RequestType.POST, __r);
+			
+			// Only the 200 range is valid for okay responses
+			int code = result.code();
+			if (!(code >= 200 && code < 300))
+			{
+				this._badresultcount.getAndIncrement();
+				
+				// Emit errors for failed requests
+				_LOGGER.error(() -> "Recv: " + result + " " +
+					__debugBody(result));
+			}
+			
+			// Debug log successful requests
+			else
+				_LOGGER.debug(() -> "Recv: " + result + " " +
+					__debugBody(result));
+		}
+		
+		// Failed to write to the server
+		catch (RemoteException e)
+		{
+			_LOGGER.error("Could not sent request to server.", e);
+			
+			this._badresultcount.getAndIncrement();
+		}
+	}
+	
+	/**
+	 * Shows string representation of the body.
+	 *
+	 * @param __b The body to decode.
+	 * @return The string result.
+	 * @since 2018/02/24
+	 */
+	private static final String __debugBody(RemoteBody __b)
+	{
+		try
+		{
+			String rv = __b.bodyAsString();
+			if (rv.indexOf('\0') >= 0)
+				return "BINARY DATA";
+			return rv;
+		}
+		catch (Throwable t)
+		{
+			return "Could not decode!";
+		}
 	}
 }
